♻️ Extract location.go
Modify getLocation method to function
